Revert "Move FSHook/PackageIndexHook/SubprocessHook to standard provider (#42…"
This reverts commit 61d1dbbc7feb9728da125dc00ad05314758036eb.
