+++ airflow/providers/standard/provider.yaml
@@ -50,10 +50,3 @@ sensors:
       - airflow.providers.standard.sensors.time_delta
       - airflow.providers.standard.sensors.time
       - airflow.providers.standard.sensors.weekday
-
-hooks:
-  - integration-name: Standard
-    python-modules:
-      - airflow.providers.standard.hooks.filesystem
-      - airflow.providers.standard.hooks.package_index
-      - airflow.providers.standard.hooks.subprocess
