+++ tests/sensors/test_filesystem.py
@@ -40,7 +40,7 @@ DEFAULT_DATE = datetime(2015, 1, 1)
 @pytest.mark.skip_if_database_isolation_mode  # Test is broken in db isolation mode
 class TestFileSensor:
     def setup_method(self):
-        from airflow.providers.standard.hooks.filesystem import FSHook
+        from airflow.hooks.filesystem import FSHook
 
         hook = FSHook()
         args = {"owner": "airflow", "start_date": DEFAULT_DATE}
